backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
web3==6.11.3
supabase==2.4.3
python-dotenv==1.0.0
boto3==1.34.0
